I change how to store the bullets in linked list using stack

+++ Game.h
@@ -11,6 +11,6 @@ public:
 	void Draw();
 
 	int velocity;
-	float positionX;
-	float positionY;
+	int positionX;
+	int positionY;
 };
